feat(clinical-history): integrate patient clinical history screen

+++ app/src/main/java/com/vitaltrace/app/feature/patient/data/mapper/ClinicalHistoryMappers.kt
@@ -28,7 +28,7 @@ private fun ClinicalTreatmentDto.toDomain() = ClinicalTreatment(
 
 private fun ClinicalMeasurementDto.toDomain() = ClinicalMeasurement(
     id,
-    value,
+    value.toDouble(),
     unit,
     measuredAt,
     observation,
